almost cerntain the problem is in the write or read to file. generation seems to be fine

+++ typedefinitions.h
@@ -42,7 +42,7 @@ typedef enum {DISABLE = 0, ENABLE = !DISABLE} FunctionalState;
 typedef enum {ERROR = 0, SUCCESS = !ERROR} ErrorStatus;
 
 void cprint(void);
-void printBits(int size, void const * const ptr);//assumes little endian
+void printBits(size_t size, void const * const ptr);//assumes little endian
 unsigned char *gen_rdm_bytestream (size_t nBytes);
 void writeToFile(unsigned const char* uc8inputBytes);
